Partial revert of r364792: caroot: switch to using echo+shell glob
On stable/11, I mistakenly only tested installation of trusted certs.
When the dir is empty, the glob remains unexpanded when it gets added to
FILES.

On stable/11 (but not 12 or head), this ends up being erroneous; it kind of
looks like the glob is being expanded to a single-word empty string rather
than leaving us with an empty FILES. Regardless, this isn't worth fixing on
stable/11, so back it out.


git-svn-id: https://svn.freebsd.org/base/stable/11@364793 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f