Treat the boot loader as the same as the kernel for what's visible
The boot loader will be growing some (limited) support for some kernel
interfaces for some of the timekeeping routines to support zstd code.
Allow the declarations for them to be visible when compiling for the
boot loader, rather than treating it like a user-space environment
(which stand.h already provides to a limited degree).


git-svn-id: https://svn.freebsd.org/base/head@364926 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f