sh: Write absolute path in command -vV and type
POSIX is pretty clear that command -v, command -V and type shall write
absolute pathnames. Therefore, we need to prepend the current directory's
name to relative pathnames.

This can happen either when PATH contains a relative pathname or when the
operand contains a slash but is not an absolute pathname.


git-svn-id: https://svn.freebsd.org/base/head@365037 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f