Use kmod.opts.mk in sys/modules/tcp/Makefile to fix standalone builds.
This was included in the original commit to head that added use of
KERN_OPTS (r361638) but was left out of the MFC since kmod.opts.mk
hadn't been MFC'd at the time.  It has since been merged.

This is a direct commit to stable/11 and stable/12.

Reported by:	Eugene Grosbein <eugen@grosbein.net>


git-svn-id: https://svn.freebsd.org/base/stable/11@365469 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f