MFC 359465: Document EINTEGRITY errors for many system calls.
EINTEGRITY was previously documented as a UFS-specific error for
mount(2).  This documents EINTEGRITY as a filesystem-independent error
that may be reported by the backing store of a filesystem.

While here, document EIO as a filesystem-independent error for both
mount(2) and posix_fadvise(2).  EIO was previously only documented for
UFS for mount(2).


git-svn-id: https://svn.freebsd.org/base/stable/12@365473 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f