libc tests: dynthr_mod: fix some WARNS issues
This is being addressed as part of a side-patch I'm working on that builds
all the things with WARNS=6, instead of relying on it being supplied in just
shallow parts of the build with higher-level Makefile.inc.

Provide a prototype for mod_main and annotate the thread function argument
as unused.

MFC after:	1 week


git-svn-id: https://svn.freebsd.org/base/head@365494 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f