MFC r361750:
Restrict enabling TCP-FASTOPEN to end-points in CLOSED or LISTEN state

Enabling TCP-FASTOPEN on an end-point which is in a state other than
CLOSED or LISTEN, is a bug in the application. So it should not work.
Also the TCP code does not (and needs not to) handle this.
While there, also simplify the setting of the TF_FASTOPEN flag.


git-svn-id: https://svn.freebsd.org/base/stable/12@365561 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f