MFC r359487:
Allow the TCP backhole detection to be disabled at all, enabled only
for IPv4, enabled only for IPv6, and enabled for IPv4 and IPv6.
The current blackhole detection might classify a temporary outage as
an MTU issue and reduces permanently the MSS. Since the consequences of
such a reduction due to a misclassification are much more drastically
for IPv4 than for IPv6, allow the administrator to enable it for IPv6 only.

Manually resolve conflict for BBR, which does not exist in stable/12


git-svn-id: https://svn.freebsd.org/base/stable/12@365568 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f